Added new file with Neural Network utils
